Forced default charset to ISO-8859-1
This charset effectively copies each byte into the low eight bits of a char.  This makes it the inverse of the ReaderBackedInputStream in
resteasy - which means at the end of the day we get the bytes back again in their original form and can then interpret them as a string in
whatever charset we want.